requirements.txt
deep-translator~=1.11.4
discord.py~=2.3.2
langid~=1.1.6
numpy~=1.25.2
pytest~=7.4.0
pytest-asyncio~=0.21.1
python-dotenv~=1.0.0
opencv-python~=4.10.0.82
tqdm~=4.66.1
motor~=3.6.0
fluent.runtime~=0.4.0
argostranslate~=1.9.1
google-generativeai~=0.6.0
cryptography~=42.0.8
ytnoti~=2.1.0
ruff~=0.6.3
